Python/requirements.txt
scipy==1.2.1
opencv_python==4.0.0.21
numpy==1.16.2
tqdm==4.31.1
tensorflow_gpu==1.13.1
librosa==0.7.1
Pillow==7.2.0
tensorflow==2.2.0
